refactor: separate ha_cluster_node_options export
